refactor(sessions): retire tui-local session title cache
DSH already owns a durable projection cache (stateVersion, header
identity binding, tail replay/invalidation) — the TUI-private
$DSH_HOME/cache/pi-tui-session-titles.json file cache was a second
parallel persistence of Host-owned derived state, with its own
staleness rules (log size+mtime), and the picker's only remaining
caller of readTitleSnapshots()/full-log title folds.

- sessions.ts is now pure presentation: the file cache, log-path
  derivation (projectKey/encodePathSegment), stat/log-facts checking,
  the JSONL/zstd title folds, and the engine batch/persistence-fallback
  loaders are deleted; short-id/workspace/age/tree/match helpers stay
- titles arrive exclusively through SessionReader.projectionBatch() (the
  official projection/cache/observation ladder)
- session-direct.ts owns the structural session-query surface types it
  consumes; the retired readTitleSnapshots shape lives only in a test
  fake that pins the adapter never calls it
- session-titles.test.ts deleted (its subject no longer exists);
  sessions.test.ts keeps the pure-helper + picker-overlay contracts
